프로젝트의 전체 부분을 마무리 해주었다.
TODO: 프로젝트 설계도, ERD, 포트플리오 짜기

+++ src/main/java/com/example/BankProject/redisCompare/service/ProductService.java
@@ -84,7 +84,7 @@ public class ProductService {
 
         // 가격순으로 정렬
         List<Product> sortedProducts = productList.stream()
-                .sorted(Comparator.comparingInt(Product::getPrice))
+                .sorted(Comparator.comparingDouble(Product::getPrice))
                 .collect(Collectors.toList());
 
         return sortedProducts;
